+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 - Add Right-Click on Portrait Dice to roll initiative. (Related Issue: [#118])
+- Add setting submenu to choose which CPR Actions used when auto-populating Basic Actions Container.  (Related Issue: [#132])
 
 ## [3.0.4] - 2025-04-29
 ### Fixed